+++ src/circularDendrogram/visualOptions.js
@@ -119,4 +119,25 @@ export const visualOptions = {
     default: true,
     group: 'labels',
   },
+
+  labelStyles: {
+    type: 'text',
+    label: 'Label',
+    group: 'labels',
+    options: [
+      { label: 'Primary', value: 'labelPrimary' },
+      { label: 'Secondary', value: 'labelSecondary' },
+      { label: 'Italic', value: 'labelItalic' },
+    ],
+    default: 'labelPrimary',
+    repeatFor: 'label',
+    repeatDefault: ['labelPrimary', 'labelSecondary', 'labelItalic'],
+  },
+
+  showLabelsOutline: {
+    type: 'boolean',
+    label: 'Show outline',
+    default: false,
+    group: 'labels',
+  },
 }